cron: Correct the step logic
I thought 'step' starts when user starts the cron, but turns out it
should start at the start of a range. And step is not a timer that
guarantees the execution of an event after a certain period of time,
it's more like a syntactic sugar with which user can specify the cron
time easier. So for example:

0-59/200 * * * *

Will only execute in the first minute of every hour, because +200 is
out-of-bound.

This patch fixes the logic.

Signed-off-by: Howard Chu <howardchu95@gmail.com>